requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
celery[redis]>=5.3.0
redis>=4.5.5
python-dotenv>=1.0.0
python-multipart>=0.0.6
boto3>=1.28.0
python-jose[cryptography]>=3.3.0
Pillow>=9.5.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.16.0
accelerate>=0.20.0
trimesh>=3.21.0
numpy>=1.24.0
huggingface_hub>=0.16.0
rembg>=2.0.0
xformers>=0.0.20
